requirements.txt
Django==6.0.3
djangorestframework==3.17.1
djangorestframework-simplejwt==5.5.1
django-phonenumber-field==8.4.0
phonenumbers==9.0.27
psycopg[binary]==3.3.3
python-dotenv==1.2.2
